cleaned up files, made eventsList to store events array used by events.js
