+++ view/src/pages/events.js
@@ -1,6 +1,6 @@
 // events.js
 
-// page displaying event flyers
+// page displaying event list component
 
 import React, { Component } from "react";
 import classNames from "classnames";
@@ -16,10 +16,10 @@ import {
   TableRow,
   Paper,
 } from "@material-ui/core";
-import Events from "../components/events"
+import Events from "../components/eventList"
 import NavBar from "../components/navbar";
 import Footer from "../components/footer";
-
+import {rows} from "../components/eventList"
 
 const styles = (theme) => ({
   gridItem: {
@@ -63,7 +63,81 @@ class events extends Component {
         <div className={classNames("header", { ready: headerReady })}>
           <p className="heading">Events</p>
         </div>
-        <Events />
+        <div>
+          <Grid
+            container
+            spacing={2}
+            height="100%"
+            width="100%"
+            alignItems="stretch"
+            justify="space-evenly"
+            className={classes.gridItem}
+            style={{ "padding-top": "40px" }}
+          >
+            {/*
+					 <Grid container
+						item md={8} xs={12}
+							spacing={2}
+							height="100%"
+							width="100%"
+							alignItems="stretch"
+							justify="space-evenly"
+							className={classes.gridItem}
+							style={{padding:'40px'}}
+						>
+							{flyers.map((flyer) => (
+							<Grid item md={4} xs={12} className={classes.gridItem}>
+								<div>
+									<br/>
+									<a href={flyer.link}>
+										<img className="movingItem"  src={flyer.image} verticalAlign="center" width = "100%"/>
+									</a>
+								</div>
+							</Grid>
+							))}
+						</Grid> 
+
+              */}
+            <div align="center">
+              <h1>Upcoming Events</h1>
+            </div>
+          </Grid>
+          <TableContainer component={Paper} align="center">
+            <Table className={classes.table} aria-label="simple table">
+              <TableHead>
+                <TableRow>
+                  <TableCell align="center" width="45%"><h2></h2></TableCell>
+                  <TableCell align="center" width="55%"><h2></h2></TableCell>
+                </TableRow>
+              </TableHead>
+              <TableBody>
+                {rows.map((row) => (
+                  <TableRow key={row.event.name}>
+                    <TableCell component="th" scope="row" align="center">
+                      <br />
+                      <h4>{row.event.name}</h4>
+                      <ListSubheader>{row.event.date}</ListSubheader>
+                      {row.event.description != '' ?
+                        (<div><h5 align="left">{row.event.description}</h5><br /> </div>) : null}
+                      {row.event.link != '' && row.event.link != null
+                        ? (<div>
+                          <Button href={row.event.link} align="center" variant="contained" color="primary" position="relative" zIndex="-3000">Join Meeting</Button>
+                          <br />
+                          <br />
+                        </div>)
+                        : null
+                      }
+                    </TableCell>
+                    <TableCell component="th" scope="row" align="center">
+                      {row.flyer != '' ?
+                        (<img src={row.flyer} width="60%" />) : null}
+                    </TableCell>
+                  </TableRow>
+                ))}
+              </TableBody>
+            </Table>
+          </TableContainer>
+        </div>
         <Footer />
       </div>
     );
